+++ src/canisters/frontend/ic-asset/src/canister_api/types/asset.rs
@@ -39,7 +39,7 @@ pub struct AssetProperties {
 }
 
 /// Sets the asset with the given properties.
-#[derive(Debug, CandidType, PartialEq, Eq, PartialOrd, Ord)]
+#[derive(Debug, Clone, CandidType, PartialEq, Eq, PartialOrd, Ord)]
 pub struct SetAssetPropertiesArguments {
     pub key: String,
     pub max_age: Option<Option<u64>>,